Improve maintainability of 'early delete' in ConfigBuilder
