+++ Configuration/Applications/python/ConfigBuilder.py
@@ -2221,11 +2221,12 @@ class ConfigBuilder(object):
         # everything else
         #
         # FIXME: remove when no longer needed
-        if "RECO" in self.stepMap or "RAW2RECO" in self.stepMap:
-                self.pythonCfgCode += "\n# Add early deletion of temporary data products to reduce peak memory need\n"
-                self.pythonCfgCode += "from Configuration.StandardSequences.earlyDeleteSettings_cff import customiseEarlyDeleteForRECO\n"
-                self.pythonCfgCode += "process = customiseEarlyDeleteForRECO(process)\n"
-                self.pythonCfgCode += "# End adding early deletion\n"
+        self.pythonCfgCode += "\n# Add early deletion of temporary data products to reduce peak memory need\n"
+        self.pythonCfgCode += "from Configuration.StandardSequences.earlyDeleteSettings_cff import customiseEarlyDeleteForRECO\n"
+        self.pythonCfgCode += "process = customiseEarlyDeleteForRECO(process)\n"
+        self.pythonCfgCode += "# End adding early deletion\n"
+        from Configuration.StandardSequences.earlyDeleteSettings_cff import customiseEarlyDeleteForRECO
+        self.process = customiseEarlyDeleteForRECO(self.process)
 
 
 	# make the .io file
