+++ RecoTracker/Configuration/python/customiseEarlyDeleteForSeeding.py
@@ -2,32 +2,8 @@ import FWCore.ParameterSet.Config as cms
 
 import collections
 
-def _hasInputTagModuleLabel(pset, moduleLabel):
-    for name in pset.parameterNames_():
-        value = getattr(pset,name)
-        if isinstance(value, cms.PSet):
-            if _hasInputTagModuleLabel(value, moduleLabel):
-                return True
-        elif isinstance(value, cms.VPSet):
-            for ps in value:
-                if _hasInputTagModuleLabel(ps, moduleLabel):
-                    return True
-        elif isinstance(value, cms.VInputTag):
-            for t in value:
-                t2 = t
-                if not isinstance(t, cms.InputTag):
-                    t2 = cms.InputTag(t2)
-                if t2.getModuleLabel() == moduleLabel:
-                    return True
-        elif isinstance(value, cms.InputTag):
-            if value.getModuleLabel() == moduleLabel:
-                return True
-    return False
-
-
-def customiseEarlyDeleteForSeeding(process):
+def customiseEarlyDeleteForSeeding(process, products):
     # Find the producers
-    products = collections.defaultdict(list)
     depends = collections.defaultdict(list)
 
     def _branchName(productType, moduleLabel, instanceLabel=""):
@@ -60,15 +36,7 @@ def customiseEarlyDeleteForSeeding(process):
                     ])
 
     if len(products) == 0:
-        return process
-
-    # Set process.options
-    if not hasattr(process, "options"):
-        process.options = cms.untracked.PSet()
-    if not hasattr(process.options, "canDeleteEarly"):
-        process.options.canDeleteEarly = cms.untracked.vstring()
-    for branches in products.itervalues():
-        process.options.canDeleteEarly.extend(branches)
+        return products
 
     # Resolve data dependencies
     #
@@ -86,15 +54,4 @@ def customiseEarlyDeleteForSeeding(process):
         name = keys.pop()
         _resolve(keys, name)
 
-    # Find the consumers
-    for moduleType in [process.producers_(), process.filters_(), process.analyzers_()]:
-        for name, module in moduleType.iteritems():
-            for producer, branches in products.iteritems():
-                if _hasInputTagModuleLabel(module, producer):
-                    #print "Module %s mightGet %s" % (name, str(branches))
-                    if hasattr(module, "mightGet"):
-                        module.mightGet.extend(branches)
-                    else:
-                        module.mightGet = cms.untracked.vstring(branches)
-
-    return process
+    return products
